+++ R/filters.R
@@ -2,13 +2,16 @@
 #'
 #' Filters are strings used in the `filter` argument of `lasR` stages to process only the points of
 #' interest. Filters follow the format 'Attribute condition value(s)', e.g.: `Z > 2`, `Intensity < 155`,
-#' `Classification == 2`, or `ReturnNumber == 1`. \cr
+#' `Classification == 2`, or `ReturnNumber == 1`. \cr\cr
 #' The available conditions include `>`, `<`, `>=`, `<=`, `==`, `!=`, `%in%`, `%out%`, and `%between%`.
 #' The supported attributes are any names of the attributes of the point cloud such as `X`, `Y`, `Z`, `Intensity`,
-#' `gpstime`, `UserData`, `ReturnNumber`, `ScanAngle`, `Amplitude` and so on.\cr
+#' `gpstime`, `UserData`, `ReturnNumber`, `ScanAngle`, `Amplitude` and so on.\cr\cr
+#' Valid filter strings are e.g. `Z > 2.5`, `UserData == 2`, `Classification %in% 1 2 3`, `Z %between 1 4`.
+#' Multiple conditions can be combined with `c("Z >= 1", "Z <= 4)`.\cr\cr.
 #' Note that filters never fail. If a filter references an attribute not present in the point cloud
 #' (e.g., `Intensity < 50` in a point cloud without intensity data), the attribute is treated as if it
-#' has a value of 0. This behavior can impact conditions like `Intensity < 50`.\cr
+#' has a value of 0. This behavior can impact conditions like `Intensity < 50` where all points would pass
+#' the test.\cr\cr
 #' For convenience, the most commonly used filters have corresponding helper functions that return
 #' the appropriate filter string. Points that satisfy the specified condition **are retained** for
 #' processing, while others are ignored for the current stage.
@@ -20,7 +23,7 @@
 #' @examples
 #' f <- system.file("extdata", "Topography.las", package="lasR")
 #' gnd = keep_class(c(2,9))
-#' reader_las(gnd)
+#' reader(gnd)
 #' triangulate(filter = keep_ground())
 #' rasterize(1, "max", filter = "Z > 5")
 #' @name filters